Moving files (#6)
* Moved Code to seperate file

* new file

---------
