+++ packages/flutter_tools/lib/src/isolated/resident_web_runner.dart
@@ -41,6 +41,7 @@ import '../web/file_generators/flutter_service_worker_js.dart';
 import '../web/file_generators/main_dart.dart' as main_dart;
 import '../web/web_device.dart';
 import '../web/web_runner.dart';
+import 'devfs_config.dart';
 import 'devfs_web.dart';
 
 /// Injectable factory to create a [ResidentWebRunner].
